// src/core/models/permission.model.ts
export enum AppPermission {
  RISK_CREATE = 'RISK_CREATE',
  RISK_EDIT = 'RISK_EDIT',
  RISK_DELETE = 'RISK_DELETE',
  RISK_FORCE_DELETE = 'RISK_FORCE_DELETE',
  RISK_VIEW = 'RISK_VIEW',
  RISK_REVIEW = 'RISK_REVIEW',
  RISK_FORCE_REVIEW = 'RISK_FORCE_REVIEW',
  RISK_RAISE = 'RISK_RAISE',
  RISK_OWNER = 'RISK_OWNER',
  RISK_UPLOAD = 'RISK_UPLOAD',
  RISK_SWOT = 'RISK_SWOT',
  REPORT_VIEW = 'REPORT_VIEW',
  ORGANISATION_ADMIN = 'ORGANISATION_ADMIN',
  ACTION_CREATE = 'ACTION_CREATE',
  ACTION_EDIT = 'ACTION_EDIT',
  ACTION_DELETE = 'ACTION_DELETE',
  ACTION_FORCE_DELETE = 'ACTION_FORCE_DELETE',
  ACTION_VIEW = 'ACTION_VIEW',
  ACTION_REVIEW = 'ACTION_REVIEW',
  ACTION_FORCE_REVIEW = 'ACTION_FORCE_REVIEW',
  ACTION_ASSIGNEE = 'ACTION_ASSIGNEE',
  ACTION_SEND_FOR_APPROVAL = 'ACTION_SEND_FOR_APPROVAL'
}